+++ pages/signup.js
@@ -15,7 +15,6 @@ import { OptionalTextField } from "../components/molecules/OptionalTextField";
 import { SelectField } from "../components/atoms/SelectField";
 import { CheckBox } from "../components/atoms/CheckBox";
 import { OptionalListField } from "../components/molecules/OptionalListField";
-import { maskEmail } from "../lib/utils/maskEmail";
 import Link from "next/link";
 import queryGraphQL from "../graphql/client";
 import getSignupPage from "../graphql/queries/signupQuery.graphql";
@@ -272,6 +271,7 @@ export default function Signup(props) {
     setIncomeLevel("");
     setPublicServant("");
     setAgreeToConditions("");
+    sessionStorage.removeItem("formData");
   };
 
   const handleSubmit = async (e) => {
@@ -376,30 +376,21 @@ export default function Signup(props) {
         errorsList.push(errors[error]);
       }
     } else {
-      //submit data to the api and then redirect to the thank you page
-      const response = await fetch("/api/sign-up", {
+      let response = await fetch("/api/check-email", {
         method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(formData),
+        body: formData.email,
       });
+      response = await response.json();
 
-      // if the response is good, redirect to the thankyou page
-      if (response.status === 201 || response.status === 200) {
-        sessionStorage.setItem("email", formData.email);
-        // Remove confirm email since it's no longer needed
-        delete formData["confirmEmail"];
-        let maskedEmail = maskEmail(formData.email);
-
+      if (response.message === "User exists") {
+        setGlobalErrorText(formField.errorMsg.errorRegistered);
+      } else if (response.message === "User does not exist") {
+        sessionStorage.setItem("formData", JSON.stringify(formData));
         await push({
-          pathname: "/thankyou",
-          query: { e: maskedEmail, ref: "signup" },
+          pathname: "/signup-review",
         });
-      } else if (response.status === 400) {
-        await setGlobalErrorText(formField.errorMsg.errorRegistered);
-      } else {
-        await setGlobalErrorText(formField.errorMsg.errorUnknown);
+      } else if (response.status === 500 || 400) {
+        setGlobalErrorText(formField.errorMsg.errorUnknown);
       }
     }
     // Checks if error exists
@@ -419,6 +410,27 @@ export default function Signup(props) {
     if (props.adobeAnalyticsUrl) {
       window.adobeDataLayer = window.adobeDataLayer || [];
       window.adobeDataLayer.push({ event: "pageLoad" });
+    }
+    let formData = JSON.parse(sessionStorage.getItem("formData"));
+    if (formData) {
+      setEmail(formData.email);
+      setConfirmEmail(formData.confirmEmail);
+      setYearOfBirthRange(formData.yearOfBirthRange);
+      setLanguage(formData.language);
+      setGender(formData.gender);
+      setProvince(formData.province);
+      setGenderOtherDetails(formData.genderOtherDetails);
+      setNativeStatus(formData.nativeStatus);
+      setDisability(formData.disability);
+      setDisabilityDetails(formData.disabilityDetails);
+      setMinority(formData.minority);
+      setMinorityGroup(
+        formData.minorityGroup === undefined ? [] : formData.minorityGroup
+      );
+      setMinorityGroupOther(formData.minorityGroupOther);
+      setIncomeLevel(formData.incomeLevel);
+      setPublicServant(formData.publicServant);
+      setAgreeToConditions(formData.agreeToConditions);
     }
   }, []);